+++ tests/Phake/Proxies/CallStubberProxyTest.php
@@ -63,9 +63,7 @@ class CallStubberProxyTest extends TestCase
     public function setUp(): void
     {
         $this->matcher1 = Phake::mock(Phake\Matchers\IChainableArgumentMatcher::class);
-        $this->obj      = new class implements Phake\IMock {
-            public $__PHAKE_info = null;
-        };
+        $this->obj      = Phake::mock(\stdClass::class);
         \PhakeTestUtil::setStubMapper($this->obj, Phake::mock(Phake\Stubber\StubMapper::class));
         $this->proxy    = new CallStubberProxy($this->matcher1, false);
     }